+++ pywbemtools/pywbemcli/__init__.py
@@ -23,14 +23,15 @@ and Python 3.
 from __future__ import absolute_import, print_function
 
 from ._cmd_namespace import *       # noqa: F403,F401
+from ._common import *              # noqa: F403,F401
+from ._pywbem_server import *       # noqa: F403,F401
 from ._cmd_class import *           # noqa: F403,F401
 from ._cmd_instance import *        # noqa: F403,F401
 from ._cmd_qualifier import *       # noqa: F403,F401
 from ._cmd_server import *          # noqa: F403,F401
 from ._cmd_connection import *      # noqa: F403,F401
 from ._cmd_profile import *         # noqa: F403,F401
-from ._common import *              # noqa: F403,F401
-from ._pywbem_server import *       # noqa: F403,F401
+from ._cmd_statistics import *      # noqa: F403,F401
 from ._context_obj import *         # noqa: F403,F401
 from ._connection_repository import *   # noqa: F403,F401
 from .pywbemcli import *            # noqa: F403,F401